refactor test for policy tags
